WIP moves browser-auth from lix server to client and removes dependencies
